+++ src/date/types.ts
@@ -13,6 +13,7 @@ import type {
 	MONTHS,
 	SECOND_FORMATS,
 	TIME_FORMATS,
+	UNIT_VARIANTS,
 	WESTERN_ZODIAC_SIGNS,
 	YEAR_FORMATS,
 	ZODIAC_PRESETS,
@@ -700,3 +701,9 @@ export type ChronosWithOptions = Partial<{
 	/** Milliseconds of the second, from 0 to 999. */
 	millisecond: Milliseconds;
 }>;
+
+export type $UnitLower = (typeof UNIT_VARIANTS)[keyof typeof UNIT_VARIANTS][number];
+
+export type $UnitAnyCase = Capitalize<$UnitLower> | Uppercase<$UnitLower> | $UnitLower;
+
+export type TimeWithUnit = `${number}${$UnitAnyCase}` | `${number} ${$UnitAnyCase}`;